+++ src/Validator/BetweenValidator.php
@@ -11,6 +11,8 @@
 
 namespace Sokil\Mongo\Validator;
 
+use Sokil\Mongo\Structure;
+
 /**
  * Validates value between specified
  *
@@ -19,7 +21,7 @@ namespace Sokil\Mongo\Validator;
 class BetweenValidator extends \Sokil\Mongo\Validator
 {
 
-    public function validateField(\Sokil\Mongo\Document $document, $fieldName, array $params)
+    public function validateField(Structure $document, $fieldName, array $params)
     {
         $value = $document->get($fieldName);
         if (!$value) {