+++ src/Validator/EmailValidator.php
@@ -10,6 +10,9 @@
  */
 
 namespace Sokil\Mongo\Validator;
+
+use Sokil\Mongo\Structure;
+
 /**
  * Email validator
  *
@@ -18,7 +21,7 @@ namespace Sokil\Mongo\Validator;
 class EmailValidator extends \Sokil\Mongo\Validator
 {
 
-    public function validateField(\Sokil\Mongo\Document $document, $fieldName, array $params)
+    public function validateField(Structure $document, $fieldName, array $params)
     {
         $value = $document->get($fieldName);