+++ src/Validator/GreaterValidator.php
@@ -11,10 +11,12 @@
 
 namespace Sokil\Mongo\Validator;
 
+use Sokil\Mongo\Structure;
+
 class GreaterValidator extends \Sokil\Mongo\Validator
 {
 
-    public function validateField(\Sokil\Mongo\Document $document, $fieldName, array $params)
+    public function validateField(Structure $document, $fieldName, array $params)
     {
         $value = $document->get($fieldName);
         if (!$value) {