+++ src/Validator/UrlValidator.php
@@ -11,9 +11,11 @@
 
 namespace Sokil\Mongo\Validator;
 
+use Sokil\Mongo\Structure;
+
 class UrlValidator extends \Sokil\Mongo\Validator
 {    
-    public function validateField(\Sokil\Mongo\Document $document, $fieldName, array $params)
+    public function validateField(Structure $document, $fieldName, array $params)
     {
         $value = $document->get($fieldName);